requirements.txt
django
django-bootstrap3
django-embed-video
